+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * --silent flag to run without spitting out outputs
 * Graceful exit if running with --baseline without configuring --baseline_repo
 * Pytest check that Python version matches latest CHANGELOG version
+* Add agents instructions to not change formatting of the code
 
 # 1.2.2